feat: 404 page / refactor: util components Create PageNotFoundComponent on 404 restructure generaic utility components to /components/UtilityComponents restructure Dialog files as components of navbar rather than siblings
