+++ tsequencer-frontend/src/components/UtilityComponents/DisplayErrorMessage/index.tsx
